Break progress indicator out to own component
